rename loger to logger
